Set base Tomcat image as an ARG
Parameterize the version of the Tomcat image we will base JasperReports Server on as an ARG.
Should be the version of Tomcat that is certified for the version of JasperReports Server being deployed. ie. the default ARG value is for 7.2 => tomcat:9.0.17-jre8
See the Platform Support Guide at https://community.jaspersoft.com/documentation to the certified Tomcat version.